refactor: unified title renderer 5 - unifying files
